backend/requirements.txt
numpy
pandas
matplotlib
scikit-learn
tensorflow
openpyxl
flask
Flask==2.3.2
Flask-CORS==4.0.0
python-dotenv==1.0.0
gunicorn==21.2.0
Werkzeug==2.3.6
python-dotenv